chore(deploy-sdk): rename hook modules
