+++ typescript/deploy-sdk/src/AltVMWarpModule.ts
@@ -28,7 +28,7 @@ import {
 
 import { AltVMDeployer } from './AltVMWarpDeployer.js';
 import { AltVMWarpRouteReader } from './AltVMWarpRouteReader.js';
-import { createHookWriter } from './hook/generic-hook-writer.js';
+import { createHookWriter } from './hook/hook-writer.js';
 import { createIsmWriter } from './ism/generic-ism-writer.js';
 import {
   ismConfigToArtifact,
