+++ typescript/deploy-sdk/src/index.ts
@@ -9,8 +9,8 @@ export { AltVMWarpModule } from './AltVMWarpModule.js';
 export { AltVMWarpRouteReader } from './AltVMWarpRouteReader.js';
 export { AltVMDeployer } from './AltVMWarpDeployer.js';
 export { coreModuleProvider } from './core-module.js';
-export { createHookReader } from './hook/generic-hook.js';
-export { HookWriter, createHookWriter } from './hook/generic-hook-writer.js';
+export { createHookReader } from './hook/hook-reader.js';
+export { HookWriter, createHookWriter } from './hook/hook-writer.js';
 export { createIsmReader } from './ism/generic-ism.js';
 export { IsmWriter, createIsmWriter } from './ism/generic-ism-writer.js';
 export {
